✨ Implement delete functionality for vehicles and associated refuels, updating use cases and repository methods.

+++ lib/domain/usecases/vehicle/delete_vehicle_usecase.dart
@@ -1,5 +1,6 @@
 import 'package:dartz/dartz.dart';
 import 'package:gasosa_app/core/errors/failure.dart';
+import 'package:gasosa_app/domain/repositories/refuel_repository.dart';
 import 'package:gasosa_app/domain/repositories/vehicle_repository.dart';
 import 'package:injectable/injectable.dart';
 
@@ -10,12 +11,17 @@ abstract class IDeleteVehicleUsecase {
 @Injectable(as: IDeleteVehicleUsecase)
 class DeleteVehicleUsecase implements IDeleteVehicleUsecase {
   final VehicleRepository _vehicleRepository;
+  final RefuelRepository _refuelRepository;
 
-  DeleteVehicleUsecase({required VehicleRepository vehicleRepository})
-    : _vehicleRepository = vehicleRepository;
+  DeleteVehicleUsecase({required VehicleRepository vehicleRepository, required RefuelRepository refuelRepository})
+    : _vehicleRepository = vehicleRepository,
+      _refuelRepository = refuelRepository;
 
   @override
-  Future<Either<Failure, void>> call(String id) {
+  Future<Either<Failure, void>> call(String id) async {
+    final deleteRefuels = await _refuelRepository.deleteRefuelsByVehicleId(id);
+    if (deleteRefuels is Left<Failure, void>) return deleteRefuels;
+
     return _vehicleRepository.deleteVehicle(id);
   }
 }
